Data Scrape Stream/Streamlit/application.py
# summary.py
import streamlit as st
from llama_index import LLMPredictor

# Function to generate PDF summary using Llama
def generate_pdf_summary(pdf_link):
    # Initialize Llama and load the PDF content
    try:
        predictor = LLMPredictor()
        # Assuming the content of the PDF is fetched by pdf_link
        pdf_content = fetch_pdf_content(pdf_link)  # Function to fetch PDF text
        summary = predictor.predict(pdf_content)
        return summary
    except Exception as e:
        st.error(f"Error generating summary: {e}")
        return None

# Function to fetch PDF content (add implementation for PDF fetching)
def fetch_pdf_content(pdf_link):
    # Add logic to fetch text content from the PDF link, depending on your setup
    pass
