Data Scrape Stream/pyproject.toml
[tool.poetry]
name = "data scrape stream"
version = "0.1.0"
description = ""
authors = ["SaurabhVyawahare <thelenstriangle@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9.7 || >3.9.7,<=3.12.7"
datasets = "^3.0.1"
docutils = "<0.21"
fastapi = {version = "^0.115.0", extras = ["all"]}
uvicorn = "^0.31.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pyjwt = "^2.9.0"
sqlalchemy = "^2.0.35"
pymysql = "^1.1.1"
mysql-connector-python = "^9.0.0"
python-multipart = "^0.0.12"
python-dotenv = "^1.0.1"
boto3 = "^1.35.38"
openai = "0.28.0"
streamlit = "^1.39.0"
snowflake-connector-python = "^3.12.3"
requests = "^2.32.3"
transformers = "^4.46.1"
torch = "^2.5.1"
pypdf2 = "^3.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
